requirement.txt
openai-whisper
cloudinary